Add files via upload
Added: join, leave, play, pause, resume, & stop commands to bot. Bot is not capable of playing music in voice channel using the play command and given a youtube url. I've noticed that the bot seems to break if a video of a length greater than roughly 12 mins is played. Looking into what could be the cause of the issue.  Currently you have to be in a voice channel and first have the bot join the channel using the join command. After joining you can use the play,pause,resume,stop commands. The leave command will disconnect the bot from the voice channel.